Atualiza controllers e repository e configurações no application.properties

+++ src/main/java/com/hospitalvet/backend/controller/AgendamentoController.java
@@ -15,7 +15,7 @@ public class AgendamentoController {
     private AgendamentoRepository agendamentoRepository;
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deletarAgendamento(@PathVariable Long id) {
+    public ResponseEntity<Void> deletarAgendamento(@PathVariable Integer id) {
         if (agendamentoRepository.existsById(id)) {
             agendamentoRepository.deleteById(id);
             return ResponseEntity.noContent().build();
