+++ src/main/java/com/hospitalvet/backend/repository/AgendamentoRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import com.hospitalvet.backend.model.Agendamento;
 
 @Repository
-public interface AgendamentoRepository extends JpaRepository<Agendamento, Long> {
-    List<Agendamento> findByUsuarioId(Long idUsuario); // ou int, se for o tipo correto
+public interface AgendamentoRepository extends JpaRepository<Agendamento, Integer> {
+    List<Agendamento> findByUsuarioId(Integer idUsuario); // ou int, se for o tipo correto
 }
